- Updated VirtuaNES to fix Mapper 0, 15, 16, 25, 27, 58, 62, 74, 83, 111, 163, 187, 193, 194, 199, 241, - Added Mapper 35, 156, 162, 175, 216, 220, 253, - Added ines_correct.h from FCEUX to patch the correct mappers and mirroring for ROMs with bad headers.

+++ src/cores/virtuanes/NES/Mapper/Mapper027.cpp
@@ -165,7 +165,7 @@ void	Mapper027::HSync( INT scanline )
 void	Mapper027::SaveState( LPBYTE p )
 {
 	for( INT i = 0; i < 9; i++ ) {
-		p[i] = reg[i];
+		p[i] = (BYTE)reg[i];
 	}
 	p[ 9] = irq_enable;
 	p[10] = irq_counter;
